[jsk_rosbag_tools] Split test to avoid indigo's error
