+++ jsk_rosbag_tools/tests/test_jsk_rosbag_tools.test
@@ -1,7 +1,7 @@
 <?xml version="1.0" ?>
 <launch>
 
-  <test test-name="test_bag_to_video" pkg="jsk_rosbag_tools" type="test_jsk_rosbag_tools.py"
+  <test test-name="test_jsk_rosbag_tools" pkg="jsk_rosbag_tools" type="test_jsk_rosbag_tools.py"
         time-limit="360.0" />
 
 </launch>
